Bluetooth: host: Split SSP to it's own module
Refactor the BR/EDR SSP feature out into it's own source file.
