+++ subsys/bluetooth/host/hci_core.c
@@ -2682,7 +2682,7 @@ int bt_unpair(uint8_t id, const bt_addr_le_t *addr)
 #endif /* CONFIG_BT_CONN */
 
 #if defined(CONFIG_BT_SMP) || defined(CONFIG_BT_BREDR)
-static enum bt_security_err security_err_get(uint8_t hci_err)
+enum bt_security_err bt_security_err_get(uint8_t hci_err)
 {
 	switch (hci_err) {
 	case BT_HCI_ERR_SUCCESS:
@@ -2978,344 +2978,6 @@ static void conn_complete(struct net_buf *buf)
 	cp->handle = evt->handle;
 
 	bt_hci_cmd_send_sync(BT_HCI_OP_READ_REMOTE_FEATURES, buf, NULL);
-}
-
-static void pin_code_req(struct net_buf *buf)
-{
-	struct bt_hci_evt_pin_code_req *evt = (void *)buf->data;
-	struct bt_conn *conn;
-
-	BT_DBG("");
-
-	conn = bt_conn_lookup_addr_br(&evt->bdaddr);
-	if (!conn) {
-		BT_ERR("Can't find conn for %s", bt_addr_str(&evt->bdaddr));
-		return;
-	}
-
-	bt_conn_pin_code_req(conn);
-	bt_conn_unref(conn);
-}
-
-static void link_key_notify(struct net_buf *buf)
-{
-	struct bt_hci_evt_link_key_notify *evt = (void *)buf->data;
-	struct bt_conn *conn;
-
-	conn = bt_conn_lookup_addr_br(&evt->bdaddr);
-	if (!conn) {
-		BT_ERR("Can't find conn for %s", bt_addr_str(&evt->bdaddr));
-		return;
-	}
-
-	BT_DBG("%s, link type 0x%02x", bt_addr_str(&evt->bdaddr), evt->key_type);
-
-	if (!conn->br.link_key) {
-		conn->br.link_key = bt_keys_get_link_key(&evt->bdaddr);
-	}
-	if (!conn->br.link_key) {
-		BT_ERR("Can't update keys for %s", bt_addr_str(&evt->bdaddr));
-		bt_conn_unref(conn);
-		return;
-	}
-
-	/* clear any old Link Key flags */
-	conn->br.link_key->flags = 0U;
-
-	switch (evt->key_type) {
-	case BT_LK_COMBINATION:
-		/*
-		 * Setting Combination Link Key as AUTHENTICATED means it was
-		 * successfully generated by 16 digits wide PIN code.
-		 */
-		if (atomic_test_and_clear_bit(conn->flags,
-					      BT_CONN_BR_LEGACY_SECURE)) {
-			conn->br.link_key->flags |= BT_LINK_KEY_AUTHENTICATED;
-		}
-		memcpy(conn->br.link_key->val, evt->link_key, 16);
-		break;
-	case BT_LK_AUTH_COMBINATION_P192:
-		conn->br.link_key->flags |= BT_LINK_KEY_AUTHENTICATED;
-		/* fall through */
-	case BT_LK_UNAUTH_COMBINATION_P192:
-		/* Mark no-bond so that link-key is removed on disconnection */
-		if (bt_conn_ssp_get_auth(conn) < BT_HCI_DEDICATED_BONDING) {
-			atomic_set_bit(conn->flags, BT_CONN_BR_NOBOND);
-		}
-
-		memcpy(conn->br.link_key->val, evt->link_key, 16);
-		break;
-	case BT_LK_AUTH_COMBINATION_P256:
-		conn->br.link_key->flags |= BT_LINK_KEY_AUTHENTICATED;
-		/* fall through */
-	case BT_LK_UNAUTH_COMBINATION_P256:
-		conn->br.link_key->flags |= BT_LINK_KEY_SC;
-
-		/* Mark no-bond so that link-key is removed on disconnection */
-		if (bt_conn_ssp_get_auth(conn) < BT_HCI_DEDICATED_BONDING) {
-			atomic_set_bit(conn->flags, BT_CONN_BR_NOBOND);
-		}
-
-		memcpy(conn->br.link_key->val, evt->link_key, 16);
-		break;
-	default:
-		BT_WARN("Unsupported Link Key type %u", evt->key_type);
-		(void)memset(conn->br.link_key->val, 0,
-			     sizeof(conn->br.link_key->val));
-		break;
-	}
-
-	if (IS_ENABLED(CONFIG_BT_SETTINGS) &&
-	    !atomic_test_bit(conn->flags, BT_CONN_BR_NOBOND)) {
-		bt_keys_link_key_store(conn->br.link_key);
-	}
-
-	bt_conn_unref(conn);
-}
-
-static void link_key_neg_reply(const bt_addr_t *bdaddr)
-{
-	struct bt_hci_cp_link_key_neg_reply *cp;
-	struct net_buf *buf;
-
-	BT_DBG("");
-
-	buf = bt_hci_cmd_create(BT_HCI_OP_LINK_KEY_NEG_REPLY, sizeof(*cp));
-	if (!buf) {
-		BT_ERR("Out of command buffers");
-		return;
-	}
-
-	cp = net_buf_add(buf, sizeof(*cp));
-	bt_addr_copy(&cp->bdaddr, bdaddr);
-	bt_hci_cmd_send_sync(BT_HCI_OP_LINK_KEY_NEG_REPLY, buf, NULL);
-}
-
-static void link_key_reply(const bt_addr_t *bdaddr, const uint8_t *lk)
-{
-	struct bt_hci_cp_link_key_reply *cp;
-	struct net_buf *buf;
-
-	BT_DBG("");
-
-	buf = bt_hci_cmd_create(BT_HCI_OP_LINK_KEY_REPLY, sizeof(*cp));
-	if (!buf) {
-		BT_ERR("Out of command buffers");
-		return;
-	}
-
-	cp = net_buf_add(buf, sizeof(*cp));
-	bt_addr_copy(&cp->bdaddr, bdaddr);
-	memcpy(cp->link_key, lk, 16);
-	bt_hci_cmd_send_sync(BT_HCI_OP_LINK_KEY_REPLY, buf, NULL);
-}
-
-static void link_key_req(struct net_buf *buf)
-{
-	struct bt_hci_evt_link_key_req *evt = (void *)buf->data;
-	struct bt_conn *conn;
-
-	BT_DBG("%s", bt_addr_str(&evt->bdaddr));
-
-	conn = bt_conn_lookup_addr_br(&evt->bdaddr);
-	if (!conn) {
-		BT_ERR("Can't find conn for %s", bt_addr_str(&evt->bdaddr));
-		link_key_neg_reply(&evt->bdaddr);
-		return;
-	}
-
-	if (!conn->br.link_key) {
-		conn->br.link_key = bt_keys_find_link_key(&evt->bdaddr);
-	}
-
-	if (!conn->br.link_key) {
-		link_key_neg_reply(&evt->bdaddr);
-		bt_conn_unref(conn);
-		return;
-	}
-
-	/*
-	 * Enforce regenerate by controller stronger link key since found one
-	 * in database not covers requested security level.
-	 */
-	if (!(conn->br.link_key->flags & BT_LINK_KEY_AUTHENTICATED) &&
-	    conn->required_sec_level > BT_SECURITY_L2) {
-		link_key_neg_reply(&evt->bdaddr);
-		bt_conn_unref(conn);
-		return;
-	}
-
-	link_key_reply(&evt->bdaddr, conn->br.link_key->val);
-	bt_conn_unref(conn);
-}
-
-static void io_capa_neg_reply(const bt_addr_t *bdaddr, const uint8_t reason)
-{
-	struct bt_hci_cp_io_capability_neg_reply *cp;
-	struct net_buf *resp_buf;
-
-	resp_buf = bt_hci_cmd_create(BT_HCI_OP_IO_CAPABILITY_NEG_REPLY,
-				     sizeof(*cp));
-	if (!resp_buf) {
-		BT_ERR("Out of command buffers");
-		return;
-	}
-
-	cp = net_buf_add(resp_buf, sizeof(*cp));
-	bt_addr_copy(&cp->bdaddr, bdaddr);
-	cp->reason = reason;
-	bt_hci_cmd_send_sync(BT_HCI_OP_IO_CAPABILITY_NEG_REPLY, resp_buf, NULL);
-}
-
-static void io_capa_resp(struct net_buf *buf)
-{
-	struct bt_hci_evt_io_capa_resp *evt = (void *)buf->data;
-	struct bt_conn *conn;
-
-	BT_DBG("remote %s, IOcapa 0x%02x, auth 0x%02x",
-	       bt_addr_str(&evt->bdaddr), evt->capability, evt->authentication);
-
-	if (evt->authentication > BT_HCI_GENERAL_BONDING_MITM) {
-		BT_ERR("Invalid remote authentication requirements");
-		io_capa_neg_reply(&evt->bdaddr,
-				  BT_HCI_ERR_UNSUPP_FEATURE_PARAM_VAL);
-		return;
-	}
-
-	if (evt->capability > BT_IO_NO_INPUT_OUTPUT) {
-		BT_ERR("Invalid remote io capability requirements");
-		io_capa_neg_reply(&evt->bdaddr,
-				  BT_HCI_ERR_UNSUPP_FEATURE_PARAM_VAL);
-		return;
-	}
-
-	conn = bt_conn_lookup_addr_br(&evt->bdaddr);
-	if (!conn) {
-		BT_ERR("Unable to find conn for %s", bt_addr_str(&evt->bdaddr));
-		return;
-	}
-
-	conn->br.remote_io_capa = evt->capability;
-	conn->br.remote_auth = evt->authentication;
-	atomic_set_bit(conn->flags, BT_CONN_BR_PAIRING);
-	bt_conn_unref(conn);
-}
-
-static void io_capa_req(struct net_buf *buf)
-{
-	struct bt_hci_evt_io_capa_req *evt = (void *)buf->data;
-	struct net_buf *resp_buf;
-	struct bt_conn *conn;
-	struct bt_hci_cp_io_capability_reply *cp;
-	uint8_t auth;
-
-	BT_DBG("");
-
-	conn = bt_conn_lookup_addr_br(&evt->bdaddr);
-	if (!conn) {
-		BT_ERR("Can't find conn for %s", bt_addr_str(&evt->bdaddr));
-		return;
-	}
-
-	resp_buf = bt_hci_cmd_create(BT_HCI_OP_IO_CAPABILITY_REPLY,
-				     sizeof(*cp));
-	if (!resp_buf) {
-		BT_ERR("Out of command buffers");
-		bt_conn_unref(conn);
-		return;
-	}
-
-	/*
-	 * Set authentication requirements when acting as pairing initiator to
-	 * 'dedicated bond' with MITM protection set if local IO capa
-	 * potentially allows it, and for acceptor, based on local IO capa and
-	 * remote's authentication set.
-	 */
-	if (atomic_test_bit(conn->flags, BT_CONN_BR_PAIRING_INITIATOR)) {
-		if (bt_conn_get_io_capa() != BT_IO_NO_INPUT_OUTPUT) {
-			auth = BT_HCI_DEDICATED_BONDING_MITM;
-		} else {
-			auth = BT_HCI_DEDICATED_BONDING;
-		}
-	} else {
-		auth = bt_conn_ssp_get_auth(conn);
-	}
-
-	cp = net_buf_add(resp_buf, sizeof(*cp));
-	bt_addr_copy(&cp->bdaddr, &evt->bdaddr);
-	cp->capability = bt_conn_get_io_capa();
-	cp->authentication = auth;
-	cp->oob_data = 0U;
-	bt_hci_cmd_send_sync(BT_HCI_OP_IO_CAPABILITY_REPLY, resp_buf, NULL);
-	bt_conn_unref(conn);
-}
-
-static void ssp_complete(struct net_buf *buf)
-{
-	struct bt_hci_evt_ssp_complete *evt = (void *)buf->data;
-	struct bt_conn *conn;
-
-	BT_DBG("status 0x%02x", evt->status);
-
-	conn = bt_conn_lookup_addr_br(&evt->bdaddr);
-	if (!conn) {
-		BT_ERR("Can't find conn for %s", bt_addr_str(&evt->bdaddr));
-		return;
-	}
-
-	bt_conn_ssp_auth_complete(conn, security_err_get(evt->status));
-	if (evt->status) {
-		bt_conn_disconnect(conn, BT_HCI_ERR_AUTH_FAIL);
-	}
-
-	bt_conn_unref(conn);
-}
-
-static void user_confirm_req(struct net_buf *buf)
-{
-	struct bt_hci_evt_user_confirm_req *evt = (void *)buf->data;
-	struct bt_conn *conn;
-
-	conn = bt_conn_lookup_addr_br(&evt->bdaddr);
-	if (!conn) {
-		BT_ERR("Can't find conn for %s", bt_addr_str(&evt->bdaddr));
-		return;
-	}
-
-	bt_conn_ssp_auth(conn, sys_le32_to_cpu(evt->passkey));
-	bt_conn_unref(conn);
-}
-
-static void user_passkey_notify(struct net_buf *buf)
-{
-	struct bt_hci_evt_user_passkey_notify *evt = (void *)buf->data;
-	struct bt_conn *conn;
-
-	BT_DBG("");
-
-	conn = bt_conn_lookup_addr_br(&evt->bdaddr);
-	if (!conn) {
-		BT_ERR("Can't find conn for %s", bt_addr_str(&evt->bdaddr));
-		return;
-	}
-
-	bt_conn_ssp_auth(conn, sys_le32_to_cpu(evt->passkey));
-	bt_conn_unref(conn);
-}
-
-static void user_passkey_req(struct net_buf *buf)
-{
-	struct bt_hci_evt_user_passkey_req *evt = (void *)buf->data;
-	struct bt_conn *conn;
-
-	conn = bt_conn_lookup_addr_br(&evt->bdaddr);
-	if (!conn) {
-		BT_ERR("Can't find conn for %s", bt_addr_str(&evt->bdaddr));
-		return;
-	}
-
-	bt_conn_ssp_auth(conn, 0);
-	bt_conn_unref(conn);
 }
 
 struct discovery_priv {
@@ -3629,54 +3291,6 @@ check_names:
 	discovery_results = NULL;
 	discovery_results_size = 0;
 	discovery_results_count = 0;
-}
-
-static void link_encr(const uint16_t handle)
-{
-	struct bt_hci_cp_set_conn_encrypt *encr;
-	struct net_buf *buf;
-
-	BT_DBG("");
-
-	buf = bt_hci_cmd_create(BT_HCI_OP_SET_CONN_ENCRYPT, sizeof(*encr));
-	if (!buf) {
-		BT_ERR("Out of command buffers");
-		return;
-	}
-
-	encr = net_buf_add(buf, sizeof(*encr));
-	encr->handle = sys_cpu_to_le16(handle);
-	encr->encrypt = 0x01;
-
-	bt_hci_cmd_send_sync(BT_HCI_OP_SET_CONN_ENCRYPT, buf, NULL);
-}
-
-static void auth_complete(struct net_buf *buf)
-{
-	struct bt_hci_evt_auth_complete *evt = (void *)buf->data;
-	struct bt_conn *conn;
-	uint16_t handle = sys_le16_to_cpu(evt->handle);
-
-	BT_DBG("status 0x%02x, handle %u", evt->status, handle);
-
-	conn = bt_conn_lookup_handle(handle);
-	if (!conn) {
-		BT_ERR("Can't find conn for handle %u", handle);
-		return;
-	}
-
-	if (evt->status) {
-		/*
-		 * Inform layers above HCI about non-zero authentication
-		 * status to make them able cleanup pending jobs.
-		 */
-		bt_conn_security_changed(conn, evt->status,
-					 security_err_get(evt->status));
-	} else {
-		link_encr(handle);
-	}
-
-	bt_conn_unref(conn);
 }
 
 static void read_remote_features_complete(struct net_buf *buf)
@@ -4118,7 +3732,7 @@ static void hci_encrypt_change(struct net_buf *buf)
 
 	if (evt->status) {
 		bt_conn_security_changed(conn, evt->status,
-					 security_err_get(evt->status));
+					 bt_security_err_get(evt->status));
 		bt_conn_unref(conn);
 		return;
 	}
@@ -4184,7 +3798,7 @@ static void hci_encrypt_key_refresh_complete(struct net_buf *buf)
 
 	if (evt->status) {
 		bt_conn_security_changed(conn, evt->status,
-					 security_err_get(evt->status));
+					 bt_security_err_get(evt->status));
 		bt_conn_unref(conn);
 		return;
 	}
@@ -5186,23 +4800,24 @@ static const struct event_handler normal_events[] = {
 		      sizeof(struct bt_hci_evt_conn_request)),
 	EVENT_HANDLER(BT_HCI_EVT_CONN_COMPLETE, conn_complete,
 		      sizeof(struct bt_hci_evt_conn_complete)),
-	EVENT_HANDLER(BT_HCI_EVT_PIN_CODE_REQ, pin_code_req,
+	EVENT_HANDLER(BT_HCI_EVT_PIN_CODE_REQ, hci_evt_pin_code_req,
 		      sizeof(struct bt_hci_evt_pin_code_req)),
-	EVENT_HANDLER(BT_HCI_EVT_LINK_KEY_NOTIFY, link_key_notify,
+	EVENT_HANDLER(BT_HCI_EVT_LINK_KEY_NOTIFY, hci_evt_link_key_notify,
 		      sizeof(struct bt_hci_evt_link_key_notify)),
-	EVENT_HANDLER(BT_HCI_EVT_LINK_KEY_REQ, link_key_req,
+	EVENT_HANDLER(BT_HCI_EVT_LINK_KEY_REQ, hci_evt_link_key_req,
 		      sizeof(struct bt_hci_evt_link_key_req)),
-	EVENT_HANDLER(BT_HCI_EVT_IO_CAPA_RESP, io_capa_resp,
+	EVENT_HANDLER(BT_HCI_EVT_IO_CAPA_RESP, hci_evt_io_capa_resp,
 		      sizeof(struct bt_hci_evt_io_capa_resp)),
-	EVENT_HANDLER(BT_HCI_EVT_IO_CAPA_REQ, io_capa_req,
+	EVENT_HANDLER(BT_HCI_EVT_IO_CAPA_REQ, hci_evt_io_capa_req,
 		      sizeof(struct bt_hci_evt_io_capa_req)),
-	EVENT_HANDLER(BT_HCI_EVT_SSP_COMPLETE, ssp_complete,
+	EVENT_HANDLER(BT_HCI_EVT_SSP_COMPLETE, hci_evt_ssp_complete,
 		      sizeof(struct bt_hci_evt_ssp_complete)),
-	EVENT_HANDLER(BT_HCI_EVT_USER_CONFIRM_REQ, user_confirm_req,
+	EVENT_HANDLER(BT_HCI_EVT_USER_CONFIRM_REQ, hci_evt_user_confirm_req,
 		      sizeof(struct bt_hci_evt_user_confirm_req)),
-	EVENT_HANDLER(BT_HCI_EVT_USER_PASSKEY_NOTIFY, user_passkey_notify,
+	EVENT_HANDLER(BT_HCI_EVT_USER_PASSKEY_NOTIFY,
+		      hci_evt_user_passkey_notify,
 		      sizeof(struct bt_hci_evt_user_passkey_notify)),
-	EVENT_HANDLER(BT_HCI_EVT_USER_PASSKEY_REQ, user_passkey_req,
+	EVENT_HANDLER(BT_HCI_EVT_USER_PASSKEY_REQ, hci_evt_user_passkey_req,
 		      sizeof(struct bt_hci_evt_user_passkey_req)),
 	EVENT_HANDLER(BT_HCI_EVT_INQUIRY_COMPLETE, inquiry_complete,
 		      sizeof(struct bt_hci_evt_inquiry_complete)),
@@ -5215,7 +4830,7 @@ static const struct event_handler normal_events[] = {
 	EVENT_HANDLER(BT_HCI_EVT_REMOTE_NAME_REQ_COMPLETE,
 		      remote_name_request_complete,
 		      sizeof(struct bt_hci_evt_remote_name_req_complete)),
-	EVENT_HANDLER(BT_HCI_EVT_AUTH_COMPLETE, auth_complete,
+	EVENT_HANDLER(BT_HCI_EVT_AUTH_COMPLETE, hci_evt_auth_complete,
 		      sizeof(struct bt_hci_evt_auth_complete)),
 	EVENT_HANDLER(BT_HCI_EVT_REMOTE_FEATURES,
 		      read_remote_features_complete,
